[chore] Move schema types and resolvers to `graphql/`
