+++ pages/api/graphql.ts
@@ -1,6 +1,6 @@
 import { ApolloServer } from '@apollo/server'
 import { startServerAndCreateNextHandler } from '@as-integrations/next'
-import { schema } from '../../apollo/schema'
+import { schema } from '../../graphql/schema'
 
 const apolloServer = new ApolloServer({ schema })
